fix: moved am0018 to am0017
