+++ pkg/validator/am0017/pull_secret_name_test.go
@@ -1,4 +1,4 @@
-package am0018
+package am0017
 
 import (
 	"testing"
