+++ pkg/validator/register/register.go
@@ -16,5 +16,5 @@ import (
 	_ "github.com/mt-sre/addon-metadata-operator/pkg/validator/am0013"
 	_ "github.com/mt-sre/addon-metadata-operator/pkg/validator/am0015"
 	_ "github.com/mt-sre/addon-metadata-operator/pkg/validator/am0016"
-	_ "github.com/mt-sre/addon-metadata-operator/pkg/validator/am0018"
+	_ "github.com/mt-sre/addon-metadata-operator/pkg/validator/am0017"
 )
